some requests fixed
include in solver's test
delete .DS_Store

+++ src/Solver/solver.c
@@ -8,16 +8,16 @@
 //else returns 0
 int Contains(int list[], int x){
     size_t i = 0;
-    while (i < BOARDSIZE && list[i] != x)
+    while ((int)i < BOARDSIZE && list[i] != x)
         i ++;
-    return (i == BOARDSIZE) ? 0 : 1;
+    return ((int)i == BOARDSIZE) ? 0 : 1;
 }
 
 //_IsLineValid checks if the lign is valid
 int IsLineValid(int board[][BOARDSIZE]){
-    for (size_t line = 0; line < BOARDSIZE; line ++){
+    for (size_t line = 0; (int)line < BOARDSIZE; line ++){
         int numbers[BOARDSIZE] = {0};
-        for (size_t column = 0; column < BOARDSIZE; column ++){
+        for (size_t column = 0; (int)column < BOARDSIZE; column ++){
             if (Contains(numbers, board[line][column]))
                 return 0;
             else{
@@ -35,9 +35,9 @@ int IsLineValid(int board[][BOARDSIZE]){
 
 //_IsColumnValid checks if the column is valid
 int IsColumnValid(int board[][BOARDSIZE]){
-     for (size_t column = 0; column < BOARDSIZE; column ++){
+     for (size_t column = 0; (int)column < BOARDSIZE; column ++){
          int numbers[BOARDSIZE] = {0};
-         for (size_t line = 0; line < BOARDSIZE; line ++){
+         for (size_t line = 0; (int)line < BOARDSIZE; line ++){
              if (Contains(numbers, board[line][column]))
                  return 0;
              else{
@@ -57,8 +57,8 @@ int IsColumnValid(int board[][BOARDSIZE]){
 int IsSquareValid(int board[][BOARDSIZE], int line, int column){
     int numbers[BOARDSIZE] = {0};
 
-    for (size_t i = line; i < line + 3; i ++){
-        for (size_t j = column; j < column + 3; j ++){
+    for (size_t i = line; (int)i < line + 3; i ++){
+        for (size_t j = column; (int)j < column + 3; j ++){
             if (Contains(numbers, board[i][j]))
                 return 0;
             else{
@@ -79,8 +79,8 @@ int IsBoardValid(int board[][BOARDSIZE]){
     if (!IsColumnValid(board) || !IsLineValid(board))
         return 0;
 
-    for (size_t line = 0; line <= 6; line += 3){
-        for (size_t column = 0; column <= 6; column += 3){
+    for (size_t line = 0; (int)line <= 6; line += 3){
+        for (size_t column = 0; (int)column <= 6; column += 3){
             if (!IsSquareValid(board, line, column))
                 return 0;
         }
@@ -93,8 +93,8 @@ int IsBoardValid(int board[][BOARDSIZE]){
 int IsSolved(int board[][BOARDSIZE]){
     int solved = 1;
     if (IsBoardValid(board)){
-        for (size_t i = 0; i < BOARDSIZE; i ++){
-            for (size_t j = 0; j < BOARDSIZE; j ++){
+        for (size_t i = 0; (int)i < BOARDSIZE; i ++){
+            for (size_t j = 0; (int)j < BOARDSIZE; j ++){
                 if (board[i][j] == '0'){
                     solved = 0;
                 }
@@ -113,8 +113,8 @@ int IsSolved(int board[][BOARDSIZE]){
 int Solve(int board[][BOARDSIZE]){
     int solve = 0;
 
-    for (size_t i = 0; i < BOARDSIZE; i ++){
-        for (size_t j = 0; j < BOARDSIZE; j ++){
+    for (size_t i = 0; (int)i < BOARDSIZE; i ++){
+        for (size_t j = 0; (int)j < BOARDSIZE; j ++){
             if (board[i][j] == 0){
                 int notTested[] = {1, 2, 3, 4, 5, 6, 7, 8, 9};
                 int nbElements= sizeof(notTested) / sizeof(notTested[0]);
